Move response interface to types folder and add type safety to getteammodels
