+++ webapp/src/components/onboarding/LLMConfigurationForm.tsx
@@ -20,6 +20,7 @@ import {
 	ModelType,
 	ModelTypeRequirements
 } from 'struct/model';
+import { TeamModelResponse } from 'types/teammodels';
 
 interface LLMOption {
 	label: string;
@@ -45,7 +46,7 @@ const LLMConfigurationForm = () => {
 
 	const [isMounted, setIsMounted] = useState(false);
 	const [submitting, setSubmitting] = useState(false);
-	const [teamModels, setTeamModels] = useState<Response>();
+	const [teamModels, setTeamModels] = useState<TeamModelResponse>();
 
 	const [userSelectedLLMType, setUserSelectedLLMType] = useState(false);
 	const [userSelectedEmbeddingType, setUserSelectedEmbeddingType] = useState(false);
@@ -456,79 +457,3 @@ const LLMConfigurationForm = () => {
 };
 
 export default LLMConfigurationForm;
-
-interface Team {
-	_id: string;
-	ownerId: string;
-	name: string;
-	orgId: string;
-	members: string[];
-	dateCreated: string;
-	permissions: Record<string, string>;
-	llmModel: Model;
-	embeddingModel: Model;
-}
-
-interface Model {
-	_id: string;
-	orgId: string;
-	teamId: string;
-	name: string;
-	model: string;
-	embeddingLength: number;
-	modelType: string;
-	type: string;
-	config: {
-		model: string;
-		api_key: string;
-		base_url: string;
-		cohere_api_key: string;
-		groq_api_key: string;
-	};
-}
-
-interface Account {
-	_id: string;
-	name: string;
-	email: string;
-	orgs: Org[];
-	currentOrg: string;
-	currentTeam: string;
-	stripe: Stripe;
-	oauth: Record<string, unknown>;
-	permissions: string;
-	onboarded: boolean;
-	_stripe: Stripe;
-}
-
-interface Org {
-	id: string;
-	name: string;
-	ownerId: string;
-	teams: TeamSummary[];
-	permissions: Record<string, string>;
-}
-
-interface TeamSummary {
-	id: string;
-	name: string;
-	ownerId: string;
-	permissions: Record<string, string>;
-}
-
-interface Stripe {
-	stripeCustomerId: string;
-	stripePlan: string;
-	stripeAddons: {
-		users: number;
-		storage: number;
-	};
-	stripeTrial: boolean;
-	stripeEndsAt: number;
-}
-
-interface Response {
-	data: Team;
-	csrf: string;
-	account: Account;
-}
