+++ webapp/src/types/dispatchtypes.ts
@@ -1,0 +1,5 @@
+import { Dispatch, SetStateAction } from 'react';
+
+import { TeamModelResponse } from './teammodels';
+
+export type GetTeamModelsDispatch = Dispatch<SetStateAction<TeamModelResponse>>;
